Moves time related functions into system.
