+++ library/src/BUILD.bazel
@@ -46,13 +46,3 @@ cc_library(
     ],
 )
 
-
-cc_library(
-    name = "timing",
-    srcs = ["TimeStatList.cpp"],
-    hdrs = ["TimeStatList.h", "TimeStat.h"],
-    copts = DDS_CPPOPTS,
-    linkopts = DDS_LINKOPTS,
-    local_defines = DDS_LOCAL_DEFINES,
-    visibility = ["//visibility:public"],
-)
