+++ library/src/SolverIF.cpp
@@ -14,7 +14,7 @@
 #include "SolverIF.h"
 #include "Init.h"
 #include "ABsearch.h"
-#include "TimerList.h"
+#include "system/TimerList.h"
 #include "system/System.h"
 #include "system/Scheduler.h"
 #include "dump.h"
