+++ library/src/system/BUILD.bazel
@@ -11,7 +11,6 @@ cc_library(
         "//library/src/utility:constants",
         "//library/src/utility:lookup_tables",
         "//library/src/data_types:dds_types",
-    "//library/src:timing",
         "//library/src/trans_table",
         "//library/src/moves:moves",
     ],
