+++ library/src/system/TimerList.h
@@ -49,7 +49,6 @@
 #include <string>
 
 #include "TimerGroup.h"
-#include "debug.h"
 
 using namespace std;
